+++ lib/src/features/phone/phone_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:lucii/src/features/shared/fancy_button.dart';
+import 'package:lucii/src/shared/fancy_button.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-final phoneDigitsProvider = StateProvider<String>((ref) => "+2348108130598");
+final phoneDigitsProvider = StateProvider<String>((ref) => "+2348108130590");
 
 class PhoneScreen extends ConsumerWidget {
   const PhoneScreen({super.key});
